add the "models" design to access squid and agents
