--- ocean_py/agent/agent_base.py
@@ -1,8 +1,0 @@
-"""
-    Agent class to provide basic functionality for all Ocean Agents
-"""
-
-class AgentBase():
-    def __init__(self, ocean):
-        """init the the AgentBase with the ocean instance"""
-        self._ocean = ocean
