+++ ocean_py/models/model_base.py
@@ -1,0 +1,9 @@
+
+"""
+    Model class to provide basic functionality for all Ocean Agents
+"""
+
+class ModelBase():
+    def __init__(self, ocean):
+        """init the the MobelBase with the ocean instance"""
+        self._ocean = ocean
